requirements.txt
-i https://pypi.org/simple
aiosqlite==0.18.0
alembic==1.9.1
amqp==5.1.1 ; python_version >= '3.6'
anyio==3.6.2 ; python_full_version >= '3.6.2'
attrs==22.2.0 ; python_version >= '3.6'
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7 ; python_version >= '3.6'
click==8.1.3 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
coverage[toml]==7.0.3 ; python_version >= '3.7'
databases==0.6.1 ; python_version >= '3.7'
exceptiongroup==1.1.0 ; python_version < '3.11'
fastapi==0.88.0
flower==1.2.0
greenlet==2.0.1 ; python_version >= '3' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.16.3 ; python_version >= '3.7'
httpx==0.23.3
humanize==4.4.0 ; python_version >= '3.7'
idna==3.4
iniconfig==1.1.1
kombu==5.2.4 ; python_version >= '3.7'
mako==1.2.4 ; python_version >= '3.7'
markupsafe==2.1.1 ; python_version >= '3.7'
numpy==1.24.1 ; python_version >= '3.10'
ormar==0.12.0
packaging==22.0 ; python_version >= '3.7'
pandas==1.5.2
pillow==9.4.0 ; python_version >= '3.7'
pluggy==1.0.0 ; python_version >= '3.6'
prometheus-client==0.15.0 ; python_version >= '3.6'
prompt-toolkit==3.0.36 ; python_full_version >= '3.6.2'
pydantic==1.10.2 ; python_version >= '3.7'
pytest==7.2.0
pytest-asyncio==0.20.3
pytest-cov==4.0.0
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.7
reportlab==3.6.12
rfc3986[idna2008]==1.5.0
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.0 ; python_version >= '3.7'
sqlalchemy==1.4.41 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
starlette==0.22.0 ; python_version >= '3.7'
tomli==2.0.1 ; python_version < '3.11'
tornado==6.2 ; python_version >= '3.7'
typing-extensions==4.4.0 ; python_version >= '3.7'
uvicorn==0.20.0
vine==5.0.0 ; python_version >= '3.6'
wcwidth==0.2.5
